Rename color.py to colors.py
